+++ src/components/Skills/Skills.css
@@ -1,0 +1,46 @@
+.skills-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 20px;
+    justify-content: space-between;
+  }
+  
+  .skill-item {
+    width: calc(50% - 10px); /* Two items per row */
+    background-color: #f1f1f1;
+    border-radius: 10px;
+    padding: 10px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  }
+  
+  .skill-name {
+    font-size: 16px;
+    font-weight: 600;
+    margin-bottom: 10px;
+  }
+  
+  .skill-bar {
+    position: relative;
+    width: 100%;
+    height: 25px;
+    background-color: #e0e0e0;
+    border-radius: 5px;
+  }
+  
+  .skill-fill {
+    height: 100%;
+    border-radius: 5px;
+    background-color: #00BC91;
+    text-align: right;
+    padding-right: 5px;
+    line-height: 25px;
+    color: #fff;
+    font-weight: bold;
+  }
+  
+  @media (max-width: 768px) {
+    .skill-item {
+      width: 100%; /* Full width on smaller screens */
+    }
+  }
+
